feat: correct loader.ts
